+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/_directory_client.py
@@ -564,7 +564,7 @@ class ShareDirectoryClient(StorageAccountHostsMixin):
 
     @distributed_trace
     def list_handles(self, recursive=False, **kwargs):
-        # type: (bool, Any) -> ItemPaged
+        # type: (bool, Any) -> ItemPaged[Handle]
         """Lists opened handles on a directory or a file under the directory.
 
         :param bool recursive: